% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11pdftools-package.R
\docType{package}
\name{cpp11poppler-package}
\alias{cpp11poppler-package}
\title{Open Source OCR Engine}
\description{
Bindings to 'libpoppler':
A tool for extracting text, fonts, attachments and metadata from a PDF file.
Also supports high quality rendering of PDF documents into PNG, JPEG, TIFF
format, or into raw bitmap vectors for posterior processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/cpp11poppler/}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11poppler/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Jeroen Ooms \email{jeroen@berkeley.edu} (\href{https://orcid.org/0000-0002-4035-0289}{ORCID}) (Author of pdftools R package)
}

Other contributors:
\itemize{
  \item Kristian Hogsberg (Author of poppler) [copyright holder]
  \item Munk School of Global Affairs and Public Policy [funder]
}

}
