% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdf.R
\name{qpdf}
\alias{qpdf}
\alias{pdf_combine}
\alias{pdf_compress}
\alias{pdf_length}
\alias{pdf_split}
\alias{pdf_subset}
\title{qpdf utilities}
\usage{
pdf_combine(input, output = NULL, password = "")

pdf_compress(input, output = NULL, linearize = FALSE, password = "")

pdf_length(input, password = "")

pdf_split(input, output = NULL, password = "")

pdf_subset(input, pages = 1, output = NULL, password = "")
}
\description{
qpdf utilities
}
\seealso{
Other cpp11poppler: 
\code{\link{cpp11poppler}},
\code{\link{pdf_ocr_text}()},
\code{\link{rendering}}
}
\concept{cpp11poppler}
